# Backend/docs/rapidapi/openapi-unified.yaml
openapi: 3.1.0
info:
  title: Social Intelligence Unified API
  version: 1.0.0
  description: |
    Abstracted social data interface across TikTok, Instagram,
    YouTube, Twitter/X, LinkedIn, and Threads.
    
    This is a unified schema that normalizes responses from multiple
    RapidAPI providers into a consistent format.
servers:
  - url: https://api.mediaposter.local/v1
    description: Local development
...
paths:
...
  /{platform}/users/{handle}/content:
    get:
      tags: [content]
      summary: List user content (posts/reels/shorts/videos)
      operationId: listUserContent
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Handle'
        - name: content_type
          in: query
          schema:
            $ref: '#/components/schemas/ContentType'
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        '200':
          description: Paginated content list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentPage'
  /{platform}/content/{content_id}:
    get:
      tags: [content]
      summary: Get single content item by ID
      operationId: getContent
      parameters:
        - $ref: '#/components/parameters/Platform'
        - name: content_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Content item with metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentItem'
...
components:
  parameters:
    Platform:
      name: platform
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Platform'
    Handle:
      name: handle
      in: path
      required: true
      schema:
        type: string
        description: Username or handle (without @)
  schemas:
    Platform:
      type: string
      enum: [tiktok, instagram, youtube, twitter, linkedin, threads, snapchat]
    ContentType:
      type: string
      enum: [post, reel, short, video, story, thread]
...
    ContentItem:
      type: object
      required: [platform, content_id, content_type]
      properties:
        platform:
          $ref: '#/components/schemas/Platform'
        content_id:
          type: string
        content_type:
          $ref: '#/components/schemas/ContentType'
        author_handle:
          type: string
        text:
          type: string
        created_at:
          type: string
          format: date-time
        like_count:
          type: integer
        comment_count:
          type: integer
        view_count:
          type: integer
        share_count:
          type: integer
        media_urls:
          type: array
          items:
            type: string
        raw:
          type: object
    ContentPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ContentItem'
        next_cursor:
          type: string
...